fix: address review comments - spread order, dialog reset, shared util extraction
Agent-Logs-Url: https://github.com/KenSuenobu/objectified-commercial/sessions/73dd0d8e-1348-49e0-96ac-ce40513b374c
